refactor: 파일 이름 수정 (📜 Heritage2 #14)
